refactor: move ExchangeOverview to a decinated package
