+++ src/main/java/edu/ntnu/idatt2003/millions/view/exchange/ExchangeView.java
@@ -1,11 +1,9 @@
 package edu.ntnu.idatt2003.millions.view.exchange;
 
-import edu.ntnu.idatt2003.millions.controller.MainController;
-import edu.ntnu.idatt2003.millions.controller.StartController;
 import edu.ntnu.idatt2003.millions.manager.GameManager;
 import edu.ntnu.idatt2003.millions.view.component.ViewHeader;
 import edu.ntnu.idatt2003.millions.view.component.WeekBar;
-import edu.ntnu.idatt2003.millions.view.dashboard.portfolio.PortfolioView;
+import edu.ntnu.idatt2003.millions.view.exchange.overview.ExchangeOverview;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.VBox;
